Relocate Kafka component config
