+++ client/kafka/v2/kafka.go
@@ -83,27 +83,6 @@ func New(brokers []string, saramaConfig *sarama.Config) *Builder {
 	}
 }
 
-// DefaultConsumerSaramaConfig function creates a Sarama configuration with a client ID derived from host name and consumer name.
-func DefaultConsumerSaramaConfig(name string, readCommitted bool) (*sarama.Config, error) {
-	host, err := os.Hostname()
-	if err != nil {
-		return nil, errors.New("failed to get hostname")
-	}
-
-	config := sarama.NewConfig()
-	config.ClientID = fmt.Sprintf("%s-%s", host, name)
-	config.Consumer.Return.Errors = true
-	config.Version = sarama.V0_11_0_0
-	if readCommitted {
-		// from Kafka documentation:
-		// Transactions were introduced in Kafka 0.11.0 wherein applications can write to multiple topics and partitions atomically. In order for this to work, consumers reading from these partitions should be configured to only read committed data. This can be achieved by by setting the isolation.level=read_committed in the consumer's configuration.
-		// In read_committed mode, the consumer will read only those transactional messages which have been successfully committed. It will continue to read non-transactional messages as before. There is no client-side buffering in read_committed mode. Instead, the end offset of a partition for a read_committed consumer would be the offset of the first message in the partition belonging to an open transaction. This offset is known as the 'Last Stable Offset'(LSO).
-		config.Consumer.IsolationLevel = sarama.ReadCommitted
-	}
-
-	return config, nil
-}
-
 // DefaultProducerSaramaConfig creates a default Sarama configuration with idempotency enabled.
 // See also:
 // * https://pkg.go.dev/github.com/Shopify/sarama#RequiredAcks
